feat: Check owner when delete bookmark

+++ src/bookmark/bookmark.service.ts
@@ -1,10 +1,15 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import fetch from 'node-fetch';
 import cheerio from 'cheerio';
 import { BookmarkEntity } from './bookmark.entity';
 import { GroupService } from '../group/group.service';
+import { User } from '../@types/users';
 
 interface BookmarkMeta {
   title: string;
@@ -55,7 +60,17 @@ export class BookmarkService {
     }
   }
 
-  async deleteBookmark(bookmarkId: string): Promise<void> {
+  async deleteBookmark(
+    user: User,
+    bookmarkId: string,
+    groupId: string,
+  ): Promise<void> {
+    const group = await this.groupService.findById(groupId);
+
+    if (group.user.id !== user.id) {
+      throw new UnauthorizedException();
+    }
+
     await this.bookmarkRepository
       .createQueryBuilder()
       .delete()
